+++ openeogeotrellis/deploy/run_graph_locally.py
@@ -1,23 +1,16 @@
-import os
 import sys
 from pathlib import Path
-
-import requests
 
 import openeogeotrellis.deploy.local
 from openeo.internal.graph_building import as_flat_graph
 from openeo.util import ensure_dir
 
-openeogeotrellis.deploy.local.setup_environment()
-
-# Can only import after setup_environment:
-from openeogeotrellis.deploy.batch_job import run_job
-
-# Avoid IPv6, to avoid hanging on https://services.terrascope.be/catalogue//collections
-requests.packages.urllib3.util.connection.HAS_IPV6 = False
-
 
 def run_graph_locally(process_graph, output_dir):
+    openeogeotrellis.deploy.local.setup_environment()
+    # Can only import after setup_environment:
+    from openeogeotrellis.backend import JOB_METADATA_FILENAME
+    from openeogeotrellis.deploy.batch_job import run_job
     output_dir = Path(output_dir)
     process_graph = as_flat_graph(process_graph)
     if "process_graph" not in process_graph:
@@ -25,7 +18,7 @@ def run_graph_locally(process_graph, output_dir):
     run_job(
         process_graph,
         output_file=output_dir / "random_folder_name",
-        metadata_file=output_dir / "metadata.json",
+        metadata_file=output_dir / JOB_METADATA_FILENAME,
         api_version="1.0.0",
         job_dir=ensure_dir(output_dir),
         dependencies=[],
@@ -37,6 +30,9 @@ def main():
     """
     for setup.py entry_points
     """
+    if len(sys.argv) < 2:
+        print("Usage: run_graph_locally.py path/to/process_graph.json [path/to/output/]")
+        sys.exit(1)
     process_graph_path = Path(sys.argv[1])
     if len(sys.argv) > 2:
         workdir = Path(sys.argv[2])