+++ openeogeotrellis/job_tracker_v2.py
@@ -417,8 +417,7 @@ class JobTracker:
         with self._double_job_registry as double_job_registry, StatsReporter(
             name="JobTracker.update_statuses stats", report=_log.info
         ) as stats, TimingLogger("JobTracker.update_statuses", logger=_log.info):
-
-            jobs_to_track = double_job_registry.get_active_jobs(max_age=14)
+            jobs_to_track = double_job_registry.get_active_jobs(max_age=3 * 30, max_updated_ago=14)
 
             for job_info in jobs_to_track:
                 stats["collected jobs"] += 1